canopen_utils cleanup
